# gateway/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 百舸API网关
  description: |
    统一百舸各模块API规范和数据结构的中间层网关服务。

    所有接口采用RPC风格，通过 `?action=XXX` 参数区分操作。
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3000/v1
    description: 本地开发服务器
...
paths:
  /aihc/jobs/DescribeJobs:
    get:
      tags:
        - Jobs
      summary: Jobs - DescribeJobs
      description: 查询指定资源池和队列下的训练任务列表
      operationId: describeJobs
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: resourcePoolId
          in: query
          required: true
          schema:
            type: string
          description: 资源池ID
        - name: queue
          in: query
          schema:
            type: string
          description: 队列名称
        - name: status
          in: query
          schema:
            type: string
          description: 任务状态过滤
        - name: keyword
          in: query
          schema:
            type: string
          description: 搜索关键字
        - name: pageNumber
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"
  /aihc/jobs/DescribeJob:
    get:
      tags:
        - Jobs
      summary: Jobs - DescribeJob
      description: 查询指定资源池和队列下的训练任务列表
      operationId: describeJob
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: resourcePoolId
          in: query
          required: true
          schema:
            type: string
          description: 资源池ID
        - name: queue
          in: query
          schema:
            type: string
          description: 队列名称
        - name: jobId
          in: query
          schema:
            type: string
          description: 任务ID（DescribeJob时必填）
        - name: needDetail
          in: query
          schema:
            type: boolean
          description: 是否需要详情（DescribeJob时可选）
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
  /aihc/jobs/CreateJob:
    post:
      tags:
        - Jobs
      summary: Jobs - CreateJob
      description: 创建或删除训练任务
      operationId: createJob
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateJobResponse"
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateJobResponse"
  /aihc/jobs/DeleteJob:
    post:
      tags:
        - Jobs
      summary: Jobs - DeleteJob
      description: 创建或删除训练任务
      operationId: deleteJob
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteJobRequest"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteJobResponse"
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateJobResponse"
  /aihc/services/DescribeServices:
    get:
      tags:
        - Services
      summary: Services - DescribeServices
      operationId: describeServices
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: serviceId
          in: query
          schema:
            type: string
          description: 服务ID（DescribeService时必填）
        - name: pageNumber
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceListResponse"
  /aihc/services/DescribeService:
    get:
      tags:
        - Services
      summary: Services - DescribeService
      operationId: describeService
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: serviceId
          in: query
          schema:
            type: string
          description: 服务ID（DescribeService时必填）
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceDetail"
  /aihc/services/CreateService:
    post:
      tags:
        - Services
      summary: Services - CreateService
      operationId: createService
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "200":
          description: 成功
        "201":
          description: 创建成功
  /aihc/services/DeleteService:
    post:
      tags:
        - Services
      summary: Services - DeleteService
      operationId: deleteService
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteServiceRequest"
      responses:
        "200":
          description: 成功
        "201":
          description: 创建成功
...
  /aihc/jobs/StopJob:
    post:
      tags:
        - Jobs
      summary: Jobs - StopJob
      operationId: stopJob
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: 动态透传参数
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /aihc/jobs/BatchStopJobs:
    post:
      tags:
        - Jobs
      summary: Jobs - BatchStopJobs
      operationId: batchStopJobs
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: 动态透传参数
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /aihc/jobs/ModifyJob:
    post:
      tags:
        - Jobs
      summary: Jobs - ModifyJob
      operationId: modifyJob
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: 动态透传参数
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /aihc/jobs/DescribePodEvents:
    post:
      tags:
        - Jobs
      summary: Jobs - DescribePodEvents
      operationId: describePodEvents
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: 动态透传参数
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /aihc/jobs/DescribeJobEvents:
    post:
      tags:
        - Jobs
      summary: Jobs - DescribeJobEvents
      operationId: describeJobEvents
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: 动态透传参数
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /aihc/jobs/DescribeJobNodes:
    post:
      tags:
        - Jobs
      summary: Jobs - DescribeJobNodes
      operationId: describeJobNodes
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: 动态透传参数
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /aihc/jobs/DescribeJobLogs:
    post:
      tags:
        - Jobs
      summary: Jobs - DescribeJobLogs
      operationId: describeJobLogs
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: 动态透传参数
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
  /aihc/jobs/DescribeJobWebterminal:
    post:
      tags:
        - Jobs
      summary: Jobs - DescribeJobWebterminal
      operationId: describeJobWebterminal
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: 动态透传参数
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
components:
  parameters:
    RegionParam:
      name: region
      in: header
      required: true
      description: 区域代码
      schema:
        type: string
        enum:
          - bj
          - gz
          - su
          - bd
          - fwh
          - yq
        default: bd
    AkParam:
      name: ak
      in: header
      required: true
      description: 百度云 Access Key
      schema:
        type: string
    SkParam:
      name: sk
      in: header
      required: true
      description: 百度云 Secret Key
      schema:
        type: string
  schemas:
    Pagination:
      type: object
      properties:
        pageNumber:
          type: integer
        pageSize:
          type: integer
        totalCount:
          type: integer
    ComputeResources:
      type: object
      required:
        - cpu
        - memory
      properties:
        cpu:
          type: number
          description: CPU核数
        memory:
          type: number
          description: 内存(GB)
        accelerator:
          type: object
          properties:
            type:
              type: string
              description: 加速卡类型
            count:
              type: integer
              description: 加速卡数量
        sharedMemory:
          type: number
          description: 共享内存(GB)
    ResourcePoolRef:
      type: object
      required:
        - poolId
        - queue
      properties:
        poolId:
          type: string
          description: 资源池ID
        poolName:
          type: string
          description: 资源池名称
        poolType:
          type: string
          enum:
            - self-managed
            - managed
        queue:
          type: string
          description: 队列名称
    ImageConfig:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          description: 镜像地址
        source:
          type: string
          enum:
            - preset
            - custom
            - ccr
        auth:
          type: object
          properties:
            username:
              type: string
            password:
              type: string
    StorageMount:
      type: object
      required:
        - name
        - mountPath
        - storageType
        - config
      properties:
        name:
          type: string
        mountPath:
          type: string
        readOnly:
          type: boolean
        storageType:
          type: string
          enum:
            - pfs
            - bos
            - cfs
            - dataset
            - hostPath
            - emptyDir
            - cds
        config:
          type: object
    CreateJobRequest:
      type: object
      required:
        - name
        - resourcePool
        - command
        - resources
        - replicas
        - image
      properties:
        name:
          type: string
        resourcePool:
          $ref: "#/components/schemas/ResourcePoolRef"
        framework:
          type: string
          enum:
            - PyTorchJob
            - TFJob
            - MPIJob
        command:
          type: string
        resources:
          $ref: "#/components/schemas/ComputeResources"
        replicas:
          type: integer
        image:
          $ref: "#/components/schemas/ImageConfig"
        storageMounts:
          type: array
          items:
            $ref: "#/components/schemas/StorageMount"
        envs:
          type: object
          additionalProperties:
            type: string
        labels:
          type: object
          additionalProperties:
            type: string
        priority:
          type: string
          enum:
            - high
            - normal
            - low
        enableRDMA:
          type: boolean
    DeleteJobRequest:
      type: object
      required:
        - resourcePoolId
        - queue
        - jobId
      properties:
        resourcePoolId:
          type: string
        queue:
          type: string
        jobId:
          type: string
    CreateJobResponse:
      type: object
      properties:
        requestId:
          type: string
        jobId:
          type: string
        jobName:
          type: string
    DeleteJobResponse:
      type: object
      properties:
        requestId:
          type: string
        jobId:
          type: string
    JobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            type: object
        pagination:
          $ref: "#/components/schemas/Pagination"
    JobDetail:
      type: object
      properties:
        jobId:
          type: string
        jobName:
          type: string
        status:
          type: string
        resourcePool:
          $ref: "#/components/schemas/ResourcePoolRef"
        framework:
          type: string
        command:
          type: string
        resources:
          $ref: "#/components/schemas/ComputeResources"
        replicas:
          type: integer
        image:
          $ref: "#/components/schemas/ImageConfig"
        envs:
          type: object
          additionalProperties:
            type: string
        labels:
          type: object
          additionalProperties:
            type: string
        priority:
          type: string
        enableRDMA:
          type: boolean
    CreateServiceRequest:
      type: object
      properties:
        name:
          type: string
        resourcePool:
          $ref: "#/components/schemas/ResourcePoolRef"
        instanceCount:
          type: integer
        containers:
          type: array
          items:
            type: object
    DeleteServiceRequest:
      type: object
      required:
        - serviceId
      properties:
        serviceId:
          type: string
    ServiceListResponse:
      type: object
      properties:
        services:
          type: array
          items:
            type: object
        pagination:
          $ref: "#/components/schemas/Pagination"
    ServiceDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        resourcePool:
          $ref: "#/components/schemas/ResourcePoolRef"
        instanceCount:
          type: integer
        containers:
          type: array
          items:
            type: object
...